requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.16.2

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and configuration
python-dotenv==1.0.0

# AI and OpenAI integration
openai==1.3.7

# Test dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 